rename k_meter.py to sourcemeter.py
